requirements.txt
PyGithub
tabulate==0.9.0
pandas==1.5.2
pyyaml